Fbv tweaks (#577)
* no message

* no message
